fix(operator): guard podMatchesPackage against init-container count drift
The compare loop walks the actual pod's init containers while indexing the
expected slice, so an actual pod carrying more init containers than the operator
builds panics with an index-out-of-range. podMatchesPackage runs on real cluster
pods from the raw-pod path (skyhook_controller.go), and no RecoverPanic is
configured, so an admission webhook injecting an init container takes the manager
down rather than failing one reconcile. The same missing check let the opposite
case pass silently: a pod with a container missing was never compared on it and
matched. Compare lengths first, which covers both.

The panic is what the "does not match (and does not panic) when the init-container
count differs" spec was written for; it was failing.

Also move jobMatchesPackage next to podMatchesPackage in job_builder.go, with its
specs beside the podMatchesPackage ones, rather than reviving job_controller.go.
That file was deleted when Job events moved onto the global delay handler, and the
base merge resolved the delete/modify conflict by restoring it whole, bringing
back jobHandlerFunc and its "job---" routing alongside the globalDelayHandler case
that replaced it.
